Completed module 5 task 1

+++ practiceProject/jpa-database-connection/src/main/java/com/ql/jpa_database_connection/service/StudentService.java
@@ -5,7 +5,9 @@ import com.ql.jpa_database_connection.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -18,15 +20,45 @@ public class StudentService {
           return studentRepository.findAll();
     }
 
-    public Optional<Student> getStudentById(Long id){
-          return studentRepository.findById(id);
+    public Student getStudentById(Long id){
+           if(studentRepository.existsById(id)) {
+                 Optional<Student> existingStudent=studentRepository.findById(id);
+                  try{
+                     return existingStudent.get();
+                  }catch (NoSuchElementException ex){
+                     throw new NoSuchElementException(ex.getMessage());
+                  }
+           }
+             throw new RuntimeException("User not found with particular id "+id);
     }
 
     public Student saveStudent(Student student){
-          return studentRepository.save(student);
+          String email=student.getEmail();
+          if(!studentRepository.existsByEmail(email.toLowerCase())){
+              return studentRepository.save(student);
+          }else{
+              throw new RuntimeException("User already present with this particular email");
+          }
     }
 
-    public void deleteStudent(Long id){
-          studentRepository.deleteById(id);
+    public void deleteStudentById(Long id){
+         if(studentRepository.existsById(id)){
+             studentRepository.deleteById(id);
+         }else{
+             throw new RuntimeException("User not found with particular id "+id);
+         }
+    }
+
+    public Student updateStudentById(Long id,Student updatedStudent){
+              Optional<Student> ExistingStudent=studentRepository.findById(id);
+             if(ExistingStudent.isPresent()){
+                      Student student= ExistingStudent.get();
+                      student.setName(updatedStudent.getName());
+                      student.setEmail(updatedStudent.getEmail());
+                      return studentRepository.save(student);
+             }else{
+                 throw new RuntimeException("User not found with particular id "+id);
+             }
+
     }
 }
